+++ packages/adhere-ui-writingboard/ctbuild.config.js
@@ -1,0 +1,7 @@
+const path = require('path');
+
+module.exports = {
+  getConfig({ webpackConfig }) {
+    webpackConfig.entry.index = path.join(__dirname, 'src', 'test', 'index.jsx');
+  },
+};